+++ test/e2e/test_cases/static_chart_types/polar_coo_sys/coxcomb_rectangle_1dis_1con.mjs
@@ -14,13 +14,6 @@ const testSteps = [
 				title: 'Coxcomb Chart',
 				orientation: 'horizontal',
 				coordSystem: 'polar'
-			},
-			style: {
-				plot: {
-					marker: {
-						rectangleSpacing: 0
-					}
-				}
 			}
 		})
 ]