+++ test/e2e/test_cases/ww_next_steps/next_steps_Tests/38_C_L_line.mjs
@@ -297,13 +297,6 @@ const testSteps = [
 				geometry: 'rectangle',
 				legend: null,
 				split: false
-			},
-			style: {
-				plot: {
-					marker: {
-						rectangleSpacing: 0
-					}
-				}
 			}
 		})
 ]