+++ test/e2e/test_cases/ww_noFade/wNoFade_cases/1_des_pol/rectangle/09_rec_TemporalDistribution.mjs
@@ -35,11 +35,6 @@ const testSteps = [
 					orientation: 'horizontal',
 					align: 'stretch',
 					split: false
-				},
-				style: {
-					plot: {
-						marker: { rectangleSpacing: 0 }
-					}
 				}
 			},
 			{
@@ -76,11 +71,6 @@ const testSteps = [
 					orientation: 'horizontal',
 					align: 'min',
 					legend: null
-				},
-				style: {
-					plot: {
-						marker: { rectangleSpacing: null }
-					}
 				}
 			},
 			{
@@ -170,7 +160,6 @@ const testSteps = [
 				style: {
 					plot: {
 						marker: {
-							rectangleSpacing: null,
 							fillOpacity: null,
 							borderWidth: null
 						}