+++ test/e2e/test_cases/ww_samples_for_presets/without_coo_sys/601_W_R_heatmap_gradient.mjs
@@ -26,7 +26,6 @@ const testSteps = [
 			style: {
 				plot: {
 					marker: {
-						rectangleSpacing: 0,
 						colorGradient:
 							'#3d51b8 0, #6389ec 0.15, #9fbffa 0.35, #d5d7d9 0.5, #f4b096 0.65, #e36c56 0.85, #ac1727 1'
 					}